fix: exclude context from Expo Router by moving to _context
- Expo Router treats every .tsx under app/ as a route. context.tsx
only exports UserProvider and useUser; it has no default export. The router tried to
load it as a page and warned "Route ./context.tsx is missing the required default
export", which can break routing and behavior.
