+++ frontend/cu-bytes/app/home.tsx
@@ -4,8 +4,8 @@ import { View, Text, TouchableOpacity } from 'react-native';
 import { router } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 
-import { styles } from './styles/style-home';
-import { useUser } from './context';
+import { styles } from './_styles/style-home';
+import { useUser } from './_context';
 
 export default function HomeScreen() {
